add several more commands to tcp

+++ StamOrga/Data/gameplay.h
@@ -33,7 +33,7 @@ class GamePlay : public QObject
     Q_PROPERTY(QString score READ score WRITE setScore NOTIFY scoreChanged)
     Q_PROPERTY(QString timestamp READ timestamp NOTIFY timestampChanged)
     Q_PROPERTY(QString competition READ competition NOTIFY competitionChanged)
-    Q_PROPERTY(quint32 index READ index WRITE setIndex NOTIFY indexChanged)
+    Q_PROPERTY(qint32 index READ index WRITE setIndex NOTIFY indexChanged)
     Q_PROPERTY(bool timeFixed READ timeFixed WRITE setTimeFixed NOTIFY timeFixedChanged)
     Q_PROPERTY(quint8 seasonIndex READ seasonIndex WRITE setSeasonIndex NOTIFY seasonIndexChanged)
     Q_PROPERTY(quint32 event READ getEvent NOTIFY eventChanged)
@@ -41,7 +41,7 @@ public:
     explicit GamePlay(QObject* parent = 0);
 
     QString home() { return this->m_home; }
-    void setHome(const QString& home)
+    void    setHome(const QString& home)
     {
         if (this->m_home != home) {
             this->m_home = home;
@@ -50,7 +50,7 @@ public:
     }
 
     QString away() { return this->m_away; }
-    void setAway(const QString& away)
+    void    setAway(const QString& away)
     {
         if (this->m_away != away) {
             this->m_away = away;
@@ -59,7 +59,7 @@ public:
     }
 
     QString score() { return this->m_score; }
-    void setScore(const QString& score)
+    void    setScore(const QString& score)
     {
         if (this->m_score != score) {
             this->m_score = score;
@@ -102,8 +102,8 @@ public:
         }
     }
 
-    quint8 index() { return this->m_index; }
-    void setIndex(const quint8 index)
+    qint32 index() { return this->m_index; }
+    void   setIndex(const qint32 index)
     {
         if (this->m_index != index) {
             this->m_index = index;
@@ -125,7 +125,7 @@ public:
 
 
     quint8 seasonIndex() { return this->m_seasonIndex; }
-    void setSeasonIndex(const quint8 seasonIndex)
+    void   setSeasonIndex(const quint8 seasonIndex)
     {
         if (this->m_seasonIndex != seasonIndex) {
             this->m_seasonIndex = seasonIndex;
@@ -134,7 +134,7 @@ public:
     }
 
     quint32 getEvent() { return this->m_eventCount; }
-    void setEvent(quint32 event)
+    void    setEvent(quint32 event)
     {
         if (this->m_eventCount != event) {
             this->m_eventCount = event;
@@ -145,7 +145,7 @@ public:
     Q_INVOKABLE QString getCompetitionLine();
 
     Q_INVOKABLE qint16 getFreeTickets() { return this->m_freeTickets; }
-    void setFreeTickets(qint16 number)
+    void               setFreeTickets(qint16 number)
     {
         this->m_freeTickets = number;
     }
@@ -161,49 +161,49 @@ public:
     }
 
     Q_INVOKABLE qint16 getAcceptedMeetingCount() { return this->m_acceptedMeeting; }
-    void setAcceptedMeetingCount(qint16 number)
+    void               setAcceptedMeetingCount(qint16 number)
     {
         this->m_acceptedMeeting = number;
     }
 
     Q_INVOKABLE qint16 getInterestedMeetingCount() { return this->m_interestedMeeting; }
-    void setInterestedMeetingCount(qint16 number)
+    void               setInterestedMeetingCount(qint16 number)
     {
         this->m_interestedMeeting = number;
     }
 
     Q_INVOKABLE qint16 getDeclinedMeetingCount() { return this->m_declinedMeeting; }
-    void setDeclinedMeetingCount(qint16 number)
+    void               setDeclinedMeetingCount(qint16 number)
     {
         this->m_declinedMeeting = number;
     }
 
     Q_INVOKABLE qint16 getMeetingInfo() { return this->m_meetingInfo; }
-    void setMeetingInfo(qint16 number)
+    void               setMeetingInfo(qint16 number)
     {
         this->m_meetingInfo = number;
     }
 
     Q_INVOKABLE qint16 getAcceptedTripCount() { return this->m_acceptedTrip; }
-    void setAcceptedTripCount(qint16 number)
+    void               setAcceptedTripCount(qint16 number)
     {
         this->m_acceptedTrip = number;
     }
 
     Q_INVOKABLE qint16 getInterestedTripCount() { return this->m_interestedTrip; }
-    void setInterestedTripCount(qint16 number)
+    void               setInterestedTripCount(qint16 number)
     {
         this->m_interestedTrip = number;
     }
 
     Q_INVOKABLE qint16 getDeclinedTripCount() { return this->m_declinedTrip; }
-    void setDeclinedTripCount(qint16 number)
+    void               setDeclinedTripCount(qint16 number)
     {
         this->m_declinedTrip = number;
     }
 
     Q_INVOKABLE qint16 getTripInfo() { return this->m_driveInfo; }
-    void setTripInfo(qint16 number)
+    void               setTripInfo(qint16 number)
     {
         this->m_driveInfo = number;
     }
@@ -236,7 +236,7 @@ private:
     QString          m_away;
     QString          m_score;
     CompetitionIndex m_comp;
-    quint32          m_index;
+    qint32           m_index;
     quint8           m_seasonIndex;
     qint64           m_timestamp;
     bool             m_timeFixed;
